regulator: qpnp-labibb: Hack LAB/IBB regulators for continuous splash
In continuous splash usecase the regulators are left up and
running in normal state.
Due to a hardware/firmware bug, if we send commands to put it
up, a complete reconfiguration will happen due to bad vreg
state when we boot the kernel: for this, raise the use count
to 1 to avoid the issue.

This hack may be temporary.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>
Signed-off-by: duckyduckg <duckyduckg65@gmail.com>